+++ bc/utils/models.py
@@ -4,14 +4,14 @@ from django.http import HttpResponseRedirect
 from django.utils.decorators import method_decorator
 from django.utils.functional import cached_property
 
-from wagtail import blocks
-from wagtail.admin.panels import FieldPanel, MultiFieldPanel
+from wagtail.admin.panels import FieldPanel, HelpPanel, MultiFieldPanel
 from wagtail.contrib.settings.models import BaseSiteSetting, register_setting
 from wagtail.fields import RichTextField, StreamField
 from wagtail.models import Orderable, Page
 from wagtail.search import index
 from wagtail.snippets.models import register_snippet
 
+from bc.utils.blocks import LinkBlock
 from bc.utils.cache import get_default_cache_control_decorator
 from bc.utils.constants import RICH_TEXT_FEATURES
 
@@ -182,31 +182,7 @@ class CallToActionSnippet(models.Model):
         related_name="+",
     )
 
-    link = StreamField(
-        blocks.StreamBlock(
-            [
-                (
-                    "external_link",
-                    blocks.StructBlock(
-                        [("url", blocks.URLBlock()), ("title", blocks.CharBlock())],
-                        icon="link",
-                    ),
-                ),
-                (
-                    "internal_link",
-                    blocks.StructBlock(
-                        [
-                            ("page", blocks.PageChooserBlock()),
-                            ("title", blocks.CharBlock(required=False)),
-                        ],
-                        icon="link",
-                    ),
-                ),
-            ],
-            required=True,
-        ),
-        blank=True,
-    )
+    link = StreamField(LinkBlock(), blank=True)
 
     panels = [
         FieldPanel("title"),
@@ -305,6 +281,7 @@ class SystemMessagesSettings(BaseSiteSetting):
             "information."
         ),
     )
+
     body_no_search_results = RichTextField(
         "No Search Results Message",
         default="<p>No results found.</p>",
@@ -314,6 +291,13 @@ class SystemMessagesSettings(BaseSiteSetting):
             "returns no results. You can include the search terms in the message by "
             "writing {searchterms}."
         ),
+    )
+
+    search_cta_title = models.CharField("Title", blank=True, max_length=255)
+    search_cta_button = StreamField(
+        LinkBlock(max_num=1),
+        blank=True,
+        verbose_name="Button",
     )
 
     def clean_fields(self, exclude=None):
@@ -335,6 +319,19 @@ class SystemMessagesSettings(BaseSiteSetting):
                 FieldPanel("body_no_search_results"),
             ],
             "Search",
+        ),
+        MultiFieldPanel(
+            [
+                HelpPanel(
+                    "This is a call to action (CTA) shown to users on the search page, "
+                    "right below the search bar.<br>"
+                    "Fill out both the title field and the button field "
+                    "for the CTA to show."
+                ),
+                FieldPanel("search_cta_title"),
+                FieldPanel("search_cta_button"),
+            ],
+            "Search Page Call to Action",
         ),
     ]